salary module refinement

+++ src/salary/salary.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { ITeacher } from 'src/teacher/interface/teacher.interface';
@@ -8,12 +8,15 @@ import { Salary } from './salary.models';
 import { ISalary } from './interface/salary.interface';
 import { UpdateSalaryDto } from './dto/update-salary.dto';
 import { GetSalaryByDateDto } from './dto/get-salary_date.dto';
+import { IExpense } from 'src/expense/interface/expense.interface';
+import { Expense } from 'src/expense/expense.models';
 
 @Injectable()
 export class SalaryService {
   constructor(
     @InjectModel(Teacher.name) private teacherModule: Model<ITeacher>,
     @InjectModel(Salary.name) private salaryModule: Model<ISalary>,
+    @InjectModel(Expense.name) private expenseModule: Model<IExpense>,
   ) {}
 
   async addSalaryOrder(dto: AddSalaryOrder) {
@@ -66,8 +69,18 @@ export class SalaryService {
     const salary = await this.salaryModule.findById({ _id: id });
     return salary;
   }
+
   async deleteSalaryById(id: string) {
+    const salary = await this.salaryModule.findById({ _id: id });
+    if (!salary) {
+      throw new NotFoundException('Запис з такою ЗП не знайдено');
+    }
     await this.salaryModule.deleteOne({ _id: id });
-    return `Successful delete`;
+
+    const deletedExpenses = await this.expenseModule.deleteMany({
+      salaryId: salary._id,
+    });
+
+    return `Successful delete. Додатково видалено ${deletedExpenses.deletedCount} розходів.`;
   }
 }
